WIP: Begin to connect depreciation w/ FERC
New module working on a connection between the depreciation records and 
FERC steam records, via the EIA master unit list.

First step is determining the versions of connections - ie are the 
depreciation records at a smaller, same or larger granularity. This now 
happens in the WIP `connect_deprish_to_ferc1.connect` function.

Closes on Issue #52.

Also I renamed lost of things to make the module names more consistent.

Also also addressed some of Zane's concerns in connecting depreciation 
to EIA.
