+++ .gitignore
@@ -1,0 +1,4 @@
+__pycache__/*
+*/.DS_Store*
+notebooks/.ipynb*
+*.pkl.gz
